Partially working message sending via RTT
So I found with some devices you can send a request that causes the
overall RTT to increase briefly, which can be observed by another
machine.

Not that reliable, I've got to to work around ~6 bits/second by taking 5
samples per bit (bunched together, resulting in it needing to be sent a
few times for the messages to be decoded).

Tested on a Router, by:
* Monitoring latency on its plain HTTP server.
* Send bits by connecting to the HTTPS server.

The HTTPS server triggers quite a hit, allowing a threshold of
0.02 seconds on my LAN.

Signed-off-by: B Horn <b@horn.uk>